+++ packages/sdk-coverage-tests/src/generate/save.js
@@ -1,7 +1,8 @@
 const fs = require('fs')
 const path = require('path')
+const prettier = require('prettier')
 
-async function createTestFiles(tests, {outDir, ext, formatter}) {
+async function createTestFiles(tests, {outDir, ext, format}) {
   const targetDirectory = path.join(process.cwd(), outDir)
 
   fs.rmdirSync(targetDirectory, {recursive: true})
@@ -9,7 +10,7 @@ async function createTestFiles(tests, {outDir, ext, formatter}) {
 
   tests.forEach(async test => {
     const filePath = path.resolve(targetDirectory, `${test.key}${ext}`)
-    fs.writeFileSync(filePath, formatter ? await formatter(test.code) : test.code)
+    fs.writeFileSync(filePath, format ? await prettier.format(test.code, format) : test.code)
   })
 }