+++ packages/sdk-coverage-tests/src/loaders/config-loader.js
@@ -1,7 +1,10 @@
 const path = require('path')
+const {isUrl, requireUrl} = require('../common-util')
 
 function configLoader({config: configPath}) {
-  const config = require(path.join(path.resolve('.'), configPath))
+  const config = isUrl(configPath)
+    ? requireUrl(configPath)
+    : require(path.join(path.resolve('.'), configPath))
 
   if (config.outPath || config.output) {
     config.outDir = config.outPath || config.output
@@ -11,6 +14,11 @@ function configLoader({config: configPath}) {
   }
   if (config.resultPath) {
     config.resultDir = config.resultPath
+  }
+
+  if (config.extends) {
+    const baseConfig = configLoader({config: config.extends})
+    return Object.assign(baseConfig, config)
   }
 
   return config