+++ packages/sdk-coverage-tests/src/loaders/spec-emitter-loader.js
@@ -1,13 +1,10 @@
+const path = require('path')
 const {isUrl, requireUrl} = require('../common-util')
 
-async function loadSpecEmitter(path) {
-  return isUrl(path) ? requireUrl(path) : require(path)
+async function specEmitterLoader({emitter: emitterPath}) {
+  return isUrl(emitterPath)
+    ? requireUrl(emitterPath)
+    : require(path.join(path.resolve('.'), emitterPath))
 }
 
-async function specEmitterLoader({emitter: emitterPath, initializeSdk}) {
-  if (initializeSdk) return initializeSdk
-  return loadSpecEmitter(emitterPath)
-}
-
-exports.loadSpecEmitter = loadSpecEmitter
 exports.specEmitterLoader = specEmitterLoader