+++ packages/sdk-coverage-tests/src/loaders/template-loader.js
@@ -1,13 +1,9 @@
 const handlebars = require('handlebars')
 const {loadFile} = require('../common-util')
 
-async function loadTemplate(path) {
-  return loadFile(path)
-}
-
 async function templateLoader({template: templatePath, testFrameworkTemplate}) {
   if (testFrameworkTemplate) return testFrameworkTemplate
-  const template = await loadTemplate(templatePath)
+  const template = await loadFile(templatePath)
   handlebars.registerHelper('tags', (context, options) => {
     if (!context || context.length <= 0) return ''
     return ` (${context.map(options.fn).join(' ')})`
@@ -15,5 +11,4 @@ async function templateLoader({template: templatePath, testFrameworkTemplate}) {
   return handlebars.compile(template, {noEscape: true})
 }
 
-exports.loadTemplate = loadTemplate
 exports.templateLoader = templateLoader